+++ packages/central-server/app/patientPortal/patientPortalRoutes.js
@@ -1,10 +1,9 @@
 import express from 'express';
-import config from 'config';
-import { suggestions } from '@tamanu/shared/services/suggestions';
 import { authModule } from './auth';
 import { patientDataRoutes } from './patientDataRoutes';
 
 import { getSurvey, createSurveyResponse } from './handlers';
+import { suggestionRoutes } from './suggestionRoutes';
 
 export const patientPortalModule = express.Router();
 
@@ -15,16 +14,4 @@ patientPortalModule.use('/me', patientDataRoutes);
 patientPortalModule.use('/survey/:surveyId', getSurvey);
 patientPortalModule.use('/surveyResponse/:assignmentId', createSurveyResponse);
 
-const allowedSuggesters = [];
-if (Array.isArray(allowedSuggesters) && allowedSuggesters.length > 0) {
-  const allowListMiddleware = names => (req, res, next) => {
-    const firstSegment = (req.path || '').replace(/^\//, '').split('/')[0];
-    if (!names.includes(firstSegment)) {
-      res.status(404).end();
-      return;
-    }
-    next();
-  };
-
-  patientPortalModule.use('/suggestions', allowListMiddleware(allowedSuggesters), suggestions);
-}
+patientPortalModule.use('/suggestions', suggestionRoutes);
